# backend/openapi_test.json
{"openapi":"3.1.0","info":{"title":"Analytic Service","version":"1.0.0"},"paths":{"/api/v1/analytics/stats":{"get":{"tags":["analytics"],"summary":"Get Stats","operationId":"get_stats_api_v1_analytics_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/analytics/top-courses":{"get":{"tags":["analytics"],"summary":"Get Top Courses","operationId":"get_top_courses_api_v1_analytics_top_courses_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":5,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/analytics/top-searches":{"get":{"tags":["analytics"],"summary":"Get Top Searches","operationId":"get_top_searches_api_v1_analytics_top_searches_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":5,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/analytics/revenue":{"get":{"tags":["analytics"],"summary":"Get Revenue","operationId":"get_revenue_api_v1_analytics_revenue_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/analytics/completions":{"get":{"tags":["analytics"],"summary":"Get Completions","operationId":"get_completions_api_v1_analytics_completions_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/analytics/search-by-hour":{"get":{"tags":["analytics"],"summary":"Get Search By Hour","operationId":"get_search_by_hour_api_v1_analytics_search_by_hour_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"tags":["health"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}